Move HDI_PROB and LEGEND_FONT_SIZE to causalpy/constants.py and derive CI labels from HDI_PROB
- Create causalpy/constants.py as single home for shared constants
- Delete causalpy/experiments/constants.py (now unused)
- Update all imports across 10 files to use causalpy.constants directly
- Fix 5 hardcoded "94%" label strings to derive from HDI_PROB (pymc_models,
  utils, regression_kink, regression_discontinuity, prepostnegd); also fix
  missing backslash before % in prepostnegd CI string
