+++ causalpy/experiments/piecewise_its.py
@@ -26,13 +26,13 @@ from matplotlib import pyplot as plt
 from patsy import dmatrices
 from sklearn.base import RegressorMixin
 
+from causalpy.constants import HDI_PROB, LEGEND_FONT_SIZE
 from causalpy.custom_exceptions import FormulaException
-from causalpy.experiments.constants import LEGEND_FONT_SIZE
 from causalpy.plot_utils import plot_xY
 from causalpy.pymc_models import LinearRegression, PyMCModel
 from causalpy.reporting import EffectSummary
 from causalpy.transforms import ramp, step  # noqa: F401
-from causalpy.utils import HDI_PROB, round_num
+from causalpy.utils import round_num
 
 from .base import BaseExperiment
