+++ causalpy/experiments/regression_discontinuity.py
@@ -29,12 +29,11 @@ from causalpy.custom_exceptions import (
     DataException,
     FormulaException,
 )
-from causalpy.experiments.constants import LEGEND_FONT_SIZE
+from causalpy.constants import HDI_PROB, LEGEND_FONT_SIZE
 from causalpy.plot_utils import plot_xY
 from causalpy.pymc_models import LinearRegression, PyMCModel
 from causalpy.reporting import EffectSummary, _effect_summary_rd
 from causalpy.utils import (
-    HDI_PROB,
     _is_variable_dummy_coded,
     convert_to_string,
     round_num,
@@ -340,7 +339,7 @@ class RegressionDiscontinuity(BaseExperiment):
             [(1 - HDI_PROB) / 2, 1 - (1 - HDI_PROB) / 2]
         ).values
         ci = (
-            r"$CI_{94\%}$"
+            rf"$CI_{{{HDI_PROB*100:.0f}\%}}$"
             + f"[{round_num(percentiles[0], round_to)}, {round_num(percentiles[1], round_to)}]"
         )
         discon = f"""
