+++ causalpy/experiments/regression_kink.py
@@ -30,11 +30,11 @@ from causalpy.plot_utils import plot_xY
 from causalpy.pymc_models import LinearRegression, PyMCModel
 from causalpy.reporting import EffectSummary, _effect_summary_rkink
 
-from causalpy.experiments.constants import LEGEND_FONT_SIZE
+from causalpy.constants import HDI_PROB, LEGEND_FONT_SIZE
 
 from .base import BaseExperiment
 from typing import Any, Literal
-from causalpy.utils import HDI_PROB, _is_variable_dummy_coded, round_num
+from causalpy.utils import _is_variable_dummy_coded, round_num
 from causalpy.custom_exceptions import (
     DataException,
     FormulaException,
@@ -276,7 +276,7 @@ class RegressionKink(BaseExperiment):
             [(1 - HDI_PROB) / 2, 1 - (1 - HDI_PROB) / 2]
         ).values
         ci = (
-            r"$CI_{94\%}$"
+            rf"$CI_{{{HDI_PROB*100:.0f}\%}}$"
             + f"[{round_num(percentiles[0], round_to if round_to is not None else 2)}, {round_num(percentiles[1], round_to if round_to is not None else 2)}]"
         )
         grad_change = f"""
