+++ causalpy/utils.py
@@ -29,7 +29,7 @@ import xarray as xr
 if TYPE_CHECKING:
     from causalpy.experiments.synthetic_control import SyntheticControl
 
-HDI_PROB: float = 0.94
+from causalpy.constants import HDI_PROB
 
 
 def _is_variable_dummy_coded(series: pd.Series) -> bool:
@@ -112,7 +112,7 @@ def convert_to_string(x: float | xr.DataArray, round_to: int | None = 2) -> str:
             [(1 - HDI_PROB) / 2, 1 - (1 - HDI_PROB) / 2]
         ).to_numpy()
         ci = (
-            r"$CI_{94\%}$"
+            rf"$CI_{{{HDI_PROB*100:.0f}\%}}$"
             + f"[{round_num(percentiles[0], round_to)}, {round_num(percentiles[1], round_to)}]"
         )
         return f"{x.mean().to_numpy():.2f}" + ci
